WIP: Refactor unit tests
