+++ dense/Cargo.toml
@@ -20,8 +20,8 @@ rand = "0.8"
 itertools = "0.12"
 rand_distr = "0.4"
 approx = { version = "0.5", features = ["num-complex"] }
-paste = "1"
 rand_chacha = "0.3"
+paste = "1"
 lapack = "0.19.*"
 thiserror = "1.0.40"
 serde = "1.*"
