+++ dense/src/array/empty_axis.rs
@@ -232,40 +232,3 @@ where
 
     orig
 }
-
-#[cfg(test)]
-mod test {
-    use crate::traits::{Shape, Stride};
-
-    use crate::{array::empty_axis::AxisPosition, rlst_dynamic_array3};
-
-    #[test]
-    fn test_insert_axis_back() {
-        let shape = [3, 7, 6];
-        let mut arr = rlst_dynamic_array3!(f64, shape);
-        arr.fill_from_seed_equally_distributed(0);
-
-        let new_arr = arr.view().insert_empty_axis(AxisPosition::Back);
-
-        assert_eq!(new_arr.shape(), [3, 7, 6, 1]);
-
-        assert_eq!(new_arr[[1, 2, 5, 0]], arr[[1, 2, 5]]);
-
-        assert_eq!(new_arr.stride(), [1, 3, 21, 126])
-    }
-
-    #[test]
-    fn test_insert_axis_front() {
-        let shape = [3, 7, 6];
-        let mut arr = rlst_dynamic_array3!(f64, shape);
-        arr.fill_from_seed_equally_distributed(0);
-
-        let new_arr = arr.view().insert_empty_axis(AxisPosition::Front);
-
-        assert_eq!(new_arr.shape(), [1, 3, 7, 6]);
-
-        assert_eq!(new_arr[[0, 1, 2, 5]], arr[[1, 2, 5]]);
-
-        assert_eq!(new_arr.stride(), [1, 1, 3, 21])
-    }
-}
