+++ dense/src/array/slice.rs
@@ -312,78 +312,3 @@ fn compute_raw_range<const NDIM: usize>(
         1 + convert_nd_raw(end_multi_index, stride),
     )
 }
-
-#[cfg(test)]
-mod test {
-
-    use crate::traits::*;
-    use crate::{layout::convert_nd_raw, rlst_dynamic_array3};
-
-    #[test]
-    fn test_create_slice() {
-        let shape = [3, 7, 6];
-        let mut arr = rlst_dynamic_array3!(f64, shape);
-
-        arr.fill_from_seed_equally_distributed(0);
-
-        let slice = arr.view().slice(1, 2);
-
-        assert_eq!(slice[[1, 5]], arr[[1, 2, 5]]);
-
-        assert_eq!(slice.shape(), [3, 6]);
-
-        let stride_expected = [arr.stride()[0], arr.stride()[2]];
-        let stride_actual = slice.stride();
-
-        assert_eq!(stride_expected, stride_actual);
-
-        let orig_data = arr.data();
-        let slice_data = slice.data();
-
-        let orig_raw_index = convert_nd_raw([1, 2, 5], arr.stride());
-        let slice_raw_index = convert_nd_raw([1, 5], slice.stride());
-
-        assert_eq!(orig_data[orig_raw_index], slice_data[slice_raw_index]);
-        assert_eq!(slice_data[slice_raw_index], slice[[1, 5]]);
-
-        let last_raw_index = convert_nd_raw([2, 5], slice.stride());
-        assert_eq!(slice_data[last_raw_index], slice[[2, 5]]);
-    }
-
-    #[test]
-    fn test_multiple_slices() {
-        let shape = [3, 7, 6];
-        let mut arr = rlst_dynamic_array3!(f64, shape);
-        arr.fill_from_seed_equally_distributed(0);
-
-        let mut slice = arr.view_mut().slice(1, 3).slice(1, 1);
-
-        slice[[2]] = 5.0;
-
-        assert_eq!(slice.shape(), [3]);
-        assert_eq!(arr[[2, 3, 1]], 5.0);
-    }
-
-    #[test]
-    fn test_slice_of_subview() {
-        let shape = [3, 7, 6];
-        let mut arr = rlst_dynamic_array3!(f64, shape);
-        arr.fill_from_seed_equally_distributed(0);
-        let mut arr2 = rlst_dynamic_array3!(f64, shape);
-        arr2.fill_from(arr.view());
-
-        let slice = arr.into_subview([1, 2, 4], [2, 3, 2]).slice(1, 2);
-
-        assert_eq!(slice.shape(), [2, 2]);
-
-        assert_eq!(slice[[1, 0]], arr2[[2, 4, 4]]);
-
-        let slice_data = slice.data();
-        let arr_data = arr2.data();
-
-        let slice_index = convert_nd_raw([1, 0], slice.stride());
-        let array_index = convert_nd_raw([2, 4, 4], arr2.stride());
-
-        assert_eq!(slice_data[slice_index], arr_data[array_index]);
-    }
-}
