+++ dense/src/linalg/inverse.rs
@@ -18,7 +18,7 @@ use super::assert_lapack_stride;
 ///
 /// The following command computes the inverse of an array `a`. The content
 /// of `a` is replaced by the inverse.
-/// ```
+/// ```ignore
 /// # use rlst_dense::rlst_dynamic_array2;
 /// # use rlst_dense::linalg::inverse::MatrixInverse;
 /// # let mut a = rlst_dynamic_array2!(f64, [3, 3]);
@@ -89,48 +89,3 @@ impl_inverse!(f64, dgetrf, dgetri);
 impl_inverse!(f32, sgetrf, sgetri);
 impl_inverse!(c32, cgetrf, cgetri);
 impl_inverse!(c64, zgetrf, zgetri);
-
-#[cfg(test)]
-mod test {
-
-    use super::*;
-
-    use crate::assert_array_abs_diff_eq;
-    use paste::paste;
-
-    use crate::array::empty_array;
-    use crate::rlst_dynamic_array2;
-
-    macro_rules! impl_inverse_tests {
-        ($scalar:ty, $tol:expr) => {
-            paste! {
-
-                #[test]
-                fn [<test_inverse_$scalar>]() {
-                    let n = 4;
-
-                    let mut a = rlst_dynamic_array2!($scalar, [n, n]);
-                    let mut b = rlst_dynamic_array2!($scalar, [n, n]);
-
-                    let mut ident = rlst_dynamic_array2!($scalar, [n, n]);
-                    ident.set_identity();
-
-                    a.fill_from_seed_equally_distributed(0);
-                    b.fill_from(a.view());
-
-                    a.view_mut().into_inverse_alloc().unwrap();
-
-                    let actual = empty_array::<$scalar, 2>().simple_mult_into_resize(a.view(), b.view());
-
-                    assert_array_abs_diff_eq!(actual, ident, $tol);
-                }
-
-            }
-        };
-    }
-
-    impl_inverse_tests!(f64, 1E-12);
-    impl_inverse_tests!(f32, 5E-6);
-    impl_inverse_tests!(c32, 5E-6);
-    impl_inverse_tests!(c64, 1E-12);
-}
