+++ dense/src/linalg/lu.rs
@@ -455,144 +455,3 @@ impl_lu!(f64, dgetrf, dgetrs);
 impl_lu!(f32, sgetrf, sgetrs);
 impl_lu!(c64, zgetrf, zgetrs);
 impl_lu!(c32, cgetrf, cgetrs);
-
-#[cfg(test)]
-mod test {
-
-    use crate::assert_array_relative_eq;
-    use approx;
-
-    use crate::rlst_dynamic_array1;
-    use crate::rlst_dynamic_array2;
-
-    use super::*;
-    use crate::array::empty_array;
-
-    use paste::paste;
-
-    macro_rules! impl_lu_tests {
-
-        ($scalar:ty, $tol:expr) => {
-            paste! {
-                #[test]
-                fn [<test_lu_thick_$scalar>]() {
-                    let dim = [8, 20];
-                    let mut arr = rlst_dynamic_array2!($scalar, dim);
-
-                    arr.fill_from_seed_normally_distributed(0);
-                    let mut arr2 = rlst_dynamic_array2!($scalar, dim);
-                    arr2.fill_from(arr.view());
-
-                    let lu = LuDecomposition::<$scalar,_>::new(arr2).unwrap();
-
-                    let mut l_mat = empty_array::<$scalar, 2>();
-                    let mut u_mat = empty_array::<$scalar, 2>();
-                    let mut p_mat = empty_array::<$scalar, 2>();
-
-                    lu.get_l_resize(l_mat.view_mut());
-                    lu.get_u_resize(u_mat.view_mut());
-                    lu.get_p_resize(p_mat.view_mut());
-
-                    let res = crate::array::empty_array::<$scalar, 2>();
-
-                    let res =
-                        res.simple_mult_into_resize(empty_array().simple_mult_into_resize(p_mat, l_mat), u_mat);
-
-                    assert_array_relative_eq!(res, arr, $tol)
-                }
-
-                #[test]
-                fn [<test_lu_square_$scalar>]() {
-                    let dim = [12, 12];
-                    let mut arr = rlst_dynamic_array2!($scalar, dim);
-
-                    arr.fill_from_seed_normally_distributed(0);
-                    let mut arr2 = rlst_dynamic_array2!($scalar, dim);
-                    arr2.fill_from(arr.view());
-
-                    let lu = LuDecomposition::<$scalar, _>::new(arr2).unwrap();
-
-                    let mut l_mat = empty_array::<$scalar, 2>();
-                    let mut u_mat = empty_array::<$scalar, 2>();
-                    let mut p_mat = empty_array::<$scalar, 2>();
-
-                    lu.get_l_resize(l_mat.view_mut());
-                    lu.get_u_resize(u_mat.view_mut());
-                    lu.get_p_resize(p_mat.view_mut());
-
-                    let res = crate::array::empty_array::<$scalar, 2>();
-
-                    let res =
-                        res.simple_mult_into_resize(empty_array().simple_mult_into_resize(p_mat, l_mat), u_mat);
-
-                    assert_array_relative_eq!(res, arr, $tol)
-                }
-
-                #[test]
-                fn [<test_lu_solve_$scalar>]() {
-                    let dim = [12, 12];
-                    let mut arr = rlst_dynamic_array2!($scalar, dim);
-                    arr.fill_from_seed_equally_distributed(0);
-                    let mut x_actual = rlst_dynamic_array1!($scalar, [dim[0]]);
-                    let mut rhs = rlst_dynamic_array1!($scalar, [dim[0]]);
-                    x_actual.fill_from_seed_equally_distributed(1);
-                    rhs.view_mut().simple_mult_into_resize(arr.view(), x_actual.view());
-
-                    let lu = LuDecomposition::<$scalar,_>::new(arr).unwrap();
-                    lu.solve_vec(LuTrans::NoTrans, rhs.view_mut()).unwrap();
-
-                    assert_array_relative_eq!(x_actual, rhs, $tol)
-                }
-
-
-
-                #[test]
-                fn [<test_lu_thin_$scalar>]() {
-                    let dim = [12, 8];
-                    let mut arr = rlst_dynamic_array2!($scalar, dim);
-
-                    arr.fill_from_seed_normally_distributed(0);
-                    let mut arr2 = rlst_dynamic_array2!($scalar, dim);
-                    arr2.fill_from(arr.view());
-
-                    let lu = LuDecomposition::<$scalar,_>::new(arr2).unwrap();
-
-                    let mut l_mat = empty_array::<$scalar, 2>();
-                    let mut u_mat = empty_array::<$scalar, 2>();
-                    let mut p_mat = empty_array::<$scalar, 2>();
-
-                    lu.get_l_resize(l_mat.view_mut());
-                    lu.get_u_resize(u_mat.view_mut());
-                    lu.get_p_resize(p_mat.view_mut());
-
-                    let res = crate::array::empty_array::<$scalar, 2>();
-
-                    let res =
-                        res.simple_mult_into_resize(empty_array().simple_mult_into_resize(p_mat, l_mat), u_mat);
-
-                    assert_array_relative_eq!(res, arr, $tol)
-                }
-
-                #[test]
-                fn [<test_det_$scalar>]() {
-                    let dim = [2, 2];
-                    let mut arr = rlst_dynamic_array2!($scalar, dim);
-                    arr[[0, 1]] = $scalar::from_real(3.0);
-                    arr[[1, 0]] = $scalar::from_real(2.0);
-
-                    let det = LuDecomposition::<$scalar, _>::new(arr).unwrap().det();
-
-                    approx::assert_relative_eq!(det, $scalar::from_real(-6.0), epsilon=$tol);
-                }
-
-
-
-            }
-        };
-    }
-
-    impl_lu_tests!(f64, 1E-12);
-    impl_lu_tests!(f32, 1E-5);
-    impl_lu_tests!(c64, 1E-12);
-    impl_lu_tests!(c32, 1E-5);
-}
