+++ dense/src/linalg/pseudo_inverse.rs
@@ -152,80 +152,3 @@ impl_pinv!(f64);
 impl_pinv!(f32);
 impl_pinv!(c32);
 impl_pinv!(c64);
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    use crate::array::empty_array;
-    use crate::assert_array_abs_diff_eq;
-    use paste::paste;
-
-    use crate::rlst_dynamic_array2;
-
-    macro_rules! impl_pinv_tests {
-        ($scalar:ty, $tol:expr) => {
-            paste! {
-
-                #[test]
-                fn [<test_thick_pinv_$scalar>]() {
-                    let shape = [5, 10];
-                    let tol = 0.0;
-
-                    let mut mat = rlst_dynamic_array2!($scalar, shape);
-                    let mut mat2 = rlst_dynamic_array2!($scalar, shape);
-
-                    let mut pinv = rlst_dynamic_array2!($scalar, [shape[1], shape[0]]);
-                    let mut ident = rlst_dynamic_array2!($scalar, [shape[0], shape[0]]);
-                    ident.set_identity();
-
-                    mat.fill_from_seed_equally_distributed(0);
-                    mat2.fill_from(mat.view());
-
-                    mat2.into_pseudo_inverse_alloc(pinv.view_mut(), tol)
-                        .unwrap();
-
-                    let actual = if shape[0] >= shape[1] {
-                        empty_array::<$scalar, 2>().simple_mult_into_resize(pinv, mat)
-                    } else {
-                        empty_array::<$scalar, 2>().simple_mult_into_resize(mat, pinv)
-                    };
-
-                    assert_array_abs_diff_eq!(actual, ident, $tol);
-                }
-
-                #[test]
-                fn [<test_thin_pinv_$scalar>]() {
-                    let shape = [10, 5];
-                    let tol = 0.0;
-
-                    let mut mat = rlst_dynamic_array2!($scalar, shape);
-                    let mut mat2 = rlst_dynamic_array2!($scalar, shape);
-
-                    let mut pinv = rlst_dynamic_array2!($scalar, [shape[1], shape[0]]);
-                    let mut ident = rlst_dynamic_array2!($scalar, [shape[1], shape[1]]);
-                    ident.set_identity();
-
-                    mat.fill_from_seed_equally_distributed(0);
-                    mat2.fill_from(mat.view());
-
-                    mat2.into_pseudo_inverse_alloc(pinv.view_mut(), tol)
-                        .unwrap();
-
-                    let actual = if shape[0] >= shape[1] {
-                        empty_array::<$scalar, 2>().simple_mult_into_resize(pinv, mat)
-                    } else {
-                        empty_array::<$scalar, 2>().simple_mult_into_resize(mat, pinv)
-                    };
-
-                    assert_array_abs_diff_eq!(actual, ident, $tol);
-                }
-            }
-        };
-    }
-
-    impl_pinv_tests!(f32, 1E-5);
-    impl_pinv_tests!(f64, 1E-12);
-    impl_pinv_tests!(c32, 1E-5);
-    impl_pinv_tests!(c64, 1E-12);
-}
