+++ dense/src/linalg/qr.rs
@@ -656,112 +656,3 @@ implement_qr_real!(f64, dgeqp3, dormqr);
 implement_qr_real!(f32, sgeqp3, sormqr);
 implement_qr_complex!(c64, zgeqp3, zunmqr);
 implement_qr_complex!(c32, cgeqp3, cunmqr);
-
-#[cfg(test)]
-mod test {
-
-    use crate::{assert_array_abs_diff_eq, assert_array_relative_eq, traits::*};
-
-    use crate::array::empty_array;
-    use crate::rlst_dynamic_array2;
-    use paste::paste;
-
-    use super::*;
-
-    macro_rules! implement_qr_tests {
-        ($scalar:ty, $tol:expr) => {
-            paste! {
-
-            #[test]
-            pub fn [<test_thin_qr_$scalar>]() {
-                let shape = [8, 5];
-                let mut mat = rlst_dynamic_array2!($scalar, shape);
-                let mut mat2 = rlst_dynamic_array2!($scalar, shape);
-
-                mat.fill_from_seed_equally_distributed(0);
-                mat2.fill_from(mat.view());
-
-                let mut r_mat = rlst_dynamic_array2!($scalar, [5, 5]);
-                let mut q_mat = rlst_dynamic_array2!($scalar, [8, 5]);
-                let mut p_mat = rlst_dynamic_array2!($scalar, [5, 5]);
-                let mut p_trans = rlst_dynamic_array2!($scalar, [5, 5]);
-                let actual = rlst_dynamic_array2!($scalar, [8, 5]);
-                let mut ident = rlst_dynamic_array2!($scalar, [5, 5]);
-                ident.set_identity();
-
-                let qr = QrDecomposition::<$scalar,_>::new(mat).unwrap();
-
-                let _ = qr.get_r(r_mat.view_mut());
-                let _ = qr.get_q_alloc(q_mat.view_mut());
-                let _ = qr.get_p(p_mat.view_mut());
-
-                p_trans.fill_from(p_mat.transpose());
-
-                let actual = empty_array::<$scalar, 2>()
-                    .simple_mult_into_resize(actual.simple_mult_into(q_mat.view(), r_mat.view()), p_trans);
-
-                assert_array_relative_eq!(actual, mat2, $tol);
-
-                let qtq = empty_array::<$scalar, 2>().mult_into_resize(
-                    crate::types::TransMode::ConjTrans,
-                    crate::types::TransMode::NoTrans,
-                    1.0.into(),
-                    q_mat.view(),
-                    q_mat.view(),
-                    1.0.into(),
-                );
-
-                assert_array_abs_diff_eq!(qtq, ident, $tol);
-            }
-
-            #[test]
-            pub fn [<test_thick_qr_$scalar>]() {
-                let shape = [5, 8];
-                let mut mat = rlst_dynamic_array2!($scalar, shape);
-                let mut mat2 = rlst_dynamic_array2!($scalar, shape);
-
-                mat.fill_from_seed_equally_distributed(0);
-                mat2.fill_from(mat.view());
-
-                let mut r_mat = rlst_dynamic_array2!($scalar, [5, 8]);
-                let mut q_mat = rlst_dynamic_array2!($scalar, [5, 5]);
-                let mut p_mat = rlst_dynamic_array2!($scalar, [8, 8]);
-                let mut p_trans = rlst_dynamic_array2!($scalar, [8, 8]);
-                let actual = rlst_dynamic_array2!($scalar, [5, 8]);
-                let mut ident = rlst_dynamic_array2!($scalar, [5, 5]);
-                ident.set_identity();
-
-                let qr = QrDecomposition::<$scalar, _>::new(mat).unwrap();
-
-                let _ = qr.get_r(r_mat.view_mut());
-                let _ = qr.get_q_alloc(q_mat.view_mut());
-                let _ = qr.get_p(p_mat.view_mut());
-
-                p_trans.fill_from(p_mat.transpose());
-
-                let actual = empty_array::<$scalar, 2>()
-                    .simple_mult_into_resize(actual.simple_mult_into(q_mat.view(), r_mat), p_trans);
-
-                assert_array_relative_eq!(actual, mat2, $tol);
-
-                let qtq = empty_array::<$scalar, 2>().mult_into_resize(
-                    crate::types::TransMode::ConjTrans,
-                    crate::types::TransMode::NoTrans,
-                    1.0.into(),
-                    q_mat.view(),
-                    q_mat.view(),
-                    1.0.into(),
-                );
-
-                assert_array_abs_diff_eq!(qtq, ident, $tol);
-            }
-
-                    }
-        };
-    }
-
-    implement_qr_tests!(f32, 1E-6);
-    implement_qr_tests!(f64, 1E-12);
-    implement_qr_tests!(c32, 1E-6);
-    implement_qr_tests!(c64, 1E-12);
-}
