+++ dense/src/linalg/svd.rs
@@ -460,139 +460,3 @@ impl_svd_real!(f64, dgesvd);
 impl_svd_real!(f32, sgesvd);
 impl_svd_complex!(c32, cgesvd);
 impl_svd_complex!(c64, zgesvd);
-
-#[cfg(test)]
-mod test {
-
-    use super::*;
-
-    use crate::assert_array_relative_eq;
-    use approx::assert_relative_eq;
-    use paste::paste;
-
-    use crate::array::empty_array;
-    use crate::{rlst_dynamic_array1, rlst_dynamic_array2};
-
-    use crate::linalg::qr::*;
-
-    macro_rules! impl_tests {
-        ($scalar:ty, $tol:expr) => {
-            paste! {
-
-                #[test]
-                fn [<test_singular_values_$scalar>]() {
-                    [<test_singular_values_impl_$scalar>](5, 10, $tol);
-                    [<test_singular_values_impl_$scalar>](10, 5, $tol);
-                }
-
-                #[test]
-                fn [<test_svd_$scalar>]() {
-                    [<test_svd_impl_$scalar>](10, 5, SvdMode::Reduced, $tol);
-                    [<test_svd_impl_$scalar>](5, 10, SvdMode::Reduced, $tol);
-                    [<test_svd_impl_$scalar>](10, 5, SvdMode::Full, $tol);
-                    [<test_svd_impl_$scalar>](5, 10, SvdMode::Full, $tol);
-                }
-
-                fn [<test_singular_values_impl_$scalar>](m: usize, n: usize, tol: <$scalar as RlstScalar>::Real) {
-                    let k = std::cmp::min(m, n);
-                    let mut mat = rlst_dynamic_array2!($scalar, [m, m]);
-                    let mut q = rlst_dynamic_array2!($scalar, [m, m]);
-                    let mut sigma = rlst_dynamic_array2!($scalar, [m, n]);
-
-                    mat.fill_from_seed_equally_distributed(0);
-                    let qr = QrDecomposition::<$scalar,_>::new(mat).unwrap();
-                    qr.get_q_alloc(q.view_mut()).unwrap();
-
-                    for index in 0..k {
-                        sigma[[index, index]] = ((k - index) as <$scalar as RlstScalar>::Real).into();
-                    }
-
-                    let a = empty_array::<$scalar, 2>().simple_mult_into_resize(q.view(), sigma.view());
-
-                    let mut singvals = rlst_dynamic_array1!(<$scalar as RlstScalar>::Real, [k]);
-
-                    a.into_singular_values_alloc(singvals.data_mut()).unwrap();
-
-                    for index in 0..k {
-                        assert_relative_eq!(singvals[[index]], sigma[[index, index]].re(), epsilon = tol);
-                    }
-                }
-
-                fn [<test_svd_impl_$scalar>](m: usize, n: usize, mode: SvdMode, tol: <$scalar as RlstScalar>::Real) {
-                    let k = std::cmp::min(m, n);
-
-                    let mut mat_u;
-                    let mut u;
-                    let mut mat_vt;
-                    let mut vt;
-                    let mut sigma;
-
-                    match mode {
-                        SvdMode::Full => {
-                            mat_u = rlst_dynamic_array2!($scalar, [m, m]);
-                            u = rlst_dynamic_array2!($scalar, [m, m]);
-                            mat_vt = rlst_dynamic_array2!($scalar, [n, n]);
-                            vt = rlst_dynamic_array2!($scalar, [n, n]);
-                            sigma = rlst_dynamic_array2!($scalar, [m, n]);
-                        }
-                        SvdMode::Reduced => {
-                            mat_u = rlst_dynamic_array2!($scalar, [m, k]);
-                            u = rlst_dynamic_array2!($scalar, [m, k]);
-                            mat_vt = rlst_dynamic_array2!($scalar, [k, n]);
-                            vt = rlst_dynamic_array2!($scalar, [k, n]);
-                            sigma = rlst_dynamic_array2!($scalar, [k, k]);
-                        }
-                    }
-
-                    mat_u.fill_from_seed_equally_distributed(0);
-                    mat_vt.fill_from_seed_equally_distributed(1);
-
-                    let qr = QrDecomposition::<$scalar,_>::new(mat_u).unwrap();
-                    qr.get_q_alloc(u.view_mut()).unwrap();
-
-                    let qr = QrDecomposition::<$scalar,_>::new(mat_vt).unwrap();
-                    qr.get_q_alloc(vt.view_mut()).unwrap();
-
-                    for index in 0..k {
-                        sigma[[index, index]] = ((k - index) as <$scalar as RlstScalar>::Real).into();
-                    }
-
-                    let a = empty_array::<$scalar, 2>().simple_mult_into_resize(
-                        empty_array::<$scalar, 2>().simple_mult_into_resize(u.view(), sigma.view()),
-                        vt.view(),
-                    );
-
-                    let mut expected = rlst_dynamic_array2!($scalar, a.shape());
-                    expected.fill_from(a.view());
-
-                    u.set_zero();
-                    vt.set_zero();
-                    sigma.set_zero();
-
-                    let mut singvals = rlst_dynamic_array1!(<$scalar as RlstScalar>::Real, [k]);
-
-                    a.into_svd_alloc(u.view_mut(), vt.view_mut(), singvals.data_mut(), mode)
-                        .unwrap();
-
-                    for index in 0..k {
-                        sigma[[index, index]] = singvals[[index]].into();
-                    }
-
-                    let actual = empty_array::<$scalar, 2>().simple_mult_into_resize(
-                        empty_array::<$scalar, 2>().simple_mult_into_resize(u, sigma),
-                        vt,
-                    );
-
-                    assert_array_relative_eq!(expected, actual, tol);
-                }
-
-
-            }
-        };
-    }
-
-    impl_tests!(f32, 1E-5);
-    impl_tests!(f64, 1E-12);
-    impl_tests!(c32, 1E-5);
-    impl_tests!(c64, 1E-12);
-}
