+++ rlst/Cargo.toml
@@ -33,6 +33,7 @@ lapack-src = { version = "0.9", features = ["accelerate"]}
 
 [dev-dependencies]
 criterion = { version = "0.3", features = ["html_reports"] }
+paste = "1"
 
 [package.metadata.docs.rs]
 rustdoc-args = [ "--html-in-header", "katex-header.html" ]
