--- rlst/src/dense.rs
@@ -1,7 +1,0 @@
-//! Imports from rlst_dense
-
-pub use rlst_dense::array::Array;
-pub use rlst_dense::rlst_dynamic_array2;
-pub use rlst_dense::rlst_dynamic_array3;
-pub use rlst_dense::rlst_dynamic_array4;
-pub use rlst_proc_macro::rlst_static_array;
